LiverCancer_Segmentation/requirements.txt
absl-py==0.7.1
asgiref==1.1.2
asn1crypto==0.24.0
astor==0.7.1
attrs==19.1.0
autobahn==19.6.2
Automat==0.7.0
backcall==0.1.0
bleach==2.1.4
certifi==2018.4.16
cffi==1.12.3
channels==1.1.8
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.7
cycler==0.10.0
daphne==1.4.2
decorator==4.4.0
defusedxml==0.6.0
Django==2.1.9
djangorestframework==3.8.2
entrypoints==0.3
gast==0.2.0
google-pasta==0.1.7
grpcio==1.14.0
h5py==2.8.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.7
imageio==2.5.0
incremental==17.5.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
joblib==0.12.2
jsonfield2==3.0.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.8
keras-contrib==2.0.8
Keras-Preprocessing==1.1.0
keras-tqdm==2.0.1
kiwisolver==1.0.1
lmdb==0.96
mahotas==1.4.4
Markdown==2.6.11
MarkupSafe==1.1.1
matplotlib==2.2.2
mistune==0.8.4
mypy-extensions==0.3.0
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==5.7.8
numexpr==2.6.6
numpy==1.15.0
opencv-python==3.4.2.17
openslide-python==1.1.1
openslides==2.3
pandas==0.23.4
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.2.0
prometheus-client==0.7.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
py-wsi==1.1
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
pyparsing==2.2.0
PyPDF2==1.26.0
pyrsistent==0.15.2
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.3
PyYAML==3.13
pyzmq==18.0.1
qtconsole==4.5.1
requests==2.19.1
roman==3.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.1.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.11.0
sklearn==0.0
tables==3.4.4
tensorboard==1.14.0
tensorflow==1.9.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
Theano==1.0.4
tornado==6.0.2
tqdm==4.24.0
traitlets==4.3.2
Twisted==19.2.1
txaio==18.8.1
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.2
zope.interface==4.6.0
